// src/lib/global/declarations/testVars.d.ts
export type GordLvl = "abaixo" | "eutrofico" | "sobrepeso" | "obeso1" | "obeso2" | "obeso3";
export type GordLvlLab =
  | "Eutrófico"
  | "Com Baixo Peso"
  | "Com Sobrepeso (não Obeso)"
  | "Obeso Grau 1"
  | "Obeso Grau 2"
  | "Obeso Grau 3";